preverjanje veljavnosti seje

+++ Vprasanja/CSharp/delete_user.php
@@ -1,6 +1,10 @@
 <?php
 include("config.php");
 session_start();
+if (!isset($_SESSION['logged_on'])) {
+  header("location: index.php");
+  die();
+}
 $id = $_SESSION['session_id'];
 mysqli_query($db, "DELETE FROM data_csharp WHERE ID='$id'");
